fix issue when can not delete user if user had password reset token, fix issue when image is not deleted if user has been deleted

+++ backend/src/main/java/com/github/rodis00/backend/image/ImageService.java
@@ -91,6 +91,9 @@ public class ImageService {
     }
 
     public void deleteUserProfilePicture(String filePath, UserEntity user) {
+        if (filePath == null) {
+            return;
+        }
         try {
             Files.deleteIfExists(Paths.get(filePath));
             user.setProfilePicture(null);
